# docs/implementation/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pongdang API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/notifications/subscriptions": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Subscriptions",
        "operationId": "subscriptions_api_data_notifications_subscriptions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Create",
        "operationId": "create_api_data_notifications_subscriptions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/notifications/subscriptions/{subscription_id}": {
      "put": {
        "tags": [
          "notifications"
        ],
        "summary": "Update",
        "operationId": "update_api_data_notifications_subscriptions__subscription_id__put",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subscription Id"
            }
          },
          {
            "name": "expected_revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Expected Revision"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_data_notifications_subscriptions__subscription_id__delete",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subscription Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/notifications/events": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Events",
        "operationId": "events_api_data_notifications_events_get",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subscription Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/notifications/subscriptions/{subscription_id}/evaluations": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Evaluations",
        "operationId": "evaluations_api_data_notifications_subscriptions__subscription_id__evaluations_get",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subscription Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluationPage": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/EvaluationView"
            },
            "type": "array",
            "title": "Rows"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "limit",
          "offset"
        ],
        "title": "EvaluationPage"
      },
      "EvaluationView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subscription_id": {
            "type": "string",
            "title": "Subscription Id"
          },
          "subscription_revision": {
            "type": "integer",
            "title": "Subscription Revision"
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          },
          "condition_state": {
            "type": "string",
            "title": "Condition State"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subscription_id",
          "subscription_revision",
          "evaluated_at",
          "condition_state",
          "evidence"
        ],
        "title": "EvaluationView"
      },
      "EventPage": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/EventView"
            },
            "type": "array",
            "title": "Rows"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "limit",
          "offset"
        ],
        "title": "EventPage"
      },
      "EventView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "subscription_id": {
            "type": "string",
            "title": "Subscription Id"
          },
          "subscription_revision": {
            "type": "integer",
            "title": "Subscription Revision"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence"
          },
          "delivery_state": {
            "type": "string",
            "title": "Delivery State"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subscription_id",
          "subscription_revision",
          "year",
          "kind",
          "state",
          "created_at",
          "evidence",
          "delivery_state",
          "attempts",
          "last_error"
        ],
        "title": "EventView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubscriptionInput": {
        "properties": {
          "spot_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Spot Id"
          },
          "year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "Year"
          },
          "timezone": {
            "type": "string",
            "maxLength": 100,
            "title": "Timezone",
            "default": "Asia/Seoul"
          },
          "minimum_temperature_c": {
            "type": "number",
            "maximum": 60.0,
            "minimum": -5.0,
            "title": "Minimum Temperature C"
          },
          "channel": {
            "type": "string",
            "enum": [
              "in_app",
              "email"
            ],
            "title": "Channel",
            "default": "in_app"
          },
          "destination": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 254
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spot_id",
          "year",
          "minimum_temperature_c"
        ],
        "title": "SubscriptionInput"
      },
      "SubscriptionPage": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/SubscriptionView"
            },
            "type": "array",
            "title": "Rows"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "limit",
          "offset"
        ],
        "title": "SubscriptionPage"
      },
      "SubscriptionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "spot_id": {
            "type": "integer",
            "title": "Spot Id"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "minimum_temperature_c": {
            "type": "number",
            "title": "Minimum Temperature C"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "condition_state": {
            "type": "string",
            "title": "Condition State"
          },
          "last_evaluated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Evaluated At"
          },
          "delivery_configuration": {
            "type": "string",
            "title": "Delivery Configuration"
          }
        },
        "type": "object",
        "required": [
          "id",
          "spot_id",
          "year",
          "timezone",
          "minimum_temperature_c",
          "channel",
          "destination",
          "active",
          "revision",
          "updated_at",
          "condition_state",
          "last_evaluated_at",
          "delivery_configuration"
        ],
        "title": "SubscriptionView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}